directory structure, decoupled job tests
