+++ .travis.yml
@@ -9,7 +9,7 @@ install:
 
 script:
   - pylint -E aztk
-  - pytest --ignore=tests/integration_tests
+  - pytest --ignore=tests/spark/integration_tests
 
 branches:
   only:
